[breaking change] - Change the following extension methods to internal
HasValidFullFilePathSyntax
HasInvalidFullFilePathSyntax
HasValidDriveSyntax
HasValidFullDirPathSyntax
HasValidUNCPathSyntax
