+++ Testing/VelaptorTests/InternalExtensionMethodsTests.cs
@@ -30,6 +30,101 @@ namespace VelaptorTests
     /// </summary>
     public class InternalExtensionMethodsTests
     {
+        private const char CrossPlatDirSeparatorChar = '/';
+
+        #region Unit Test Data
+        // ReSharper disable HeapView.BoxingAllocation
+
+        /// <summary>
+        /// Provides unit test data for the <see cref="InternalExtensionMethods.HasValidFullFilePathSyntax"/>() method.
+        /// </summary>
+        /// <returns>The test data.</returns>
+        public static IEnumerable<object[]> HasValidFullFilePathSyntaxTestData()
+        {
+            yield return new object[] { "C:windows", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}test-file", false };
+            yield return new object[] { string.Empty, false };
+            yield return new object[] { null, false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}", false };
+            yield return new object[] { "non-path-value", false };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}", false };
+            yield return new object[] { $@"test-dir{CrossPlatDirSeparatorChar}", false };
+            yield return new object[] { @"C:\test-dir\test-file.txt", true };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}test-file.txt", true };
+        }
+
+        /// <summary>
+        /// Provides unit test data for the <see cref="InternalExtensionMethods.HasInvalidFullFilePathSyntax"/>() method.
+        /// </summary>
+        /// <returns>The test data.</returns>
+        public static IEnumerable<object[]> IsInvalidFilePathTestData()
+        {
+            yield return new object[] { $@"C:\test-dir\test-file.txt", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}test-file.txt", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}test-file", true };
+            yield return new object[] { string.Empty, true };
+            yield return new object[] { null, true };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}", true };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}", true };
+            yield return new object[] { "non-path-value", true };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}test-dir{CrossPlatDirSeparatorChar}", true };
+            yield return new object[] { $@"test-dir{CrossPlatDirSeparatorChar}", true };
+            yield return new object[] { "C:windows", true };
+        }
+
+        /// <summary>
+        /// Provides unit test data for the <see cref="InternalExtensionMethods.HasValidDriveSyntax"/>() method.
+        /// </summary>
+        /// <returns>The test data.</returns>
+        public static IEnumerable<object[]> ContainsValidDriveTestData()
+        {
+            yield return new object[] { string.Empty, false };
+            yield return new object[] { null, false };
+            yield return new object[] { "windows", false };
+            yield return new object[] { ":", false };
+            yield return new object[] { "C", false };
+            yield return new object[] { ":C", false };
+            yield return new object[] { "1:", false };
+            yield return new object[] { "windowsC:system32", false };
+            yield return new object[] { $@"C:\Windows\System32", true };
+            yield return new object[] { @"C:windows", true };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}Windows{CrossPlatDirSeparatorChar}System32", true };
+        }
+
+        /// <summary>
+        /// Provides unit test data for the <see cref="InternalExtensionMethods.HasValidFullDirPathSyntax"/>() method.
+        /// </summary>
+        /// <returns>The test data.</returns>
+        public static IEnumerable<object[]> IsFullyQualifiedDirPathTestData()
+        {
+            yield return new object[] { string.Empty, false };
+            yield return new object[] { null, false };
+            yield return new object[] { $@"\Windows\System32", false };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}Windows{CrossPlatDirSeparatorChar}System32", false };
+            yield return new object[] { "C:Windows", false };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}WindowsC:", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}Windows{CrossPlatDirSeparatorChar}System32{CrossPlatDirSeparatorChar}fake-file.txt", false };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}Windows{CrossPlatDirSeparatorChar}System32", true };
+            yield return new object[] { $@"C:{CrossPlatDirSeparatorChar}Windows{CrossPlatDirSeparatorChar}System32{CrossPlatDirSeparatorChar}", true };
+        }
+
+        /// <summary>
+        /// Provides unit test data for the <see cref="InternalExtensionMethods.HasValidUNCPathSyntax"/>() method.
+        /// </summary>
+        /// <returns>The test data.</returns>
+        public static IEnumerable<object[]> IsUNCPathTestData()
+        {
+            yield return new object[] { string.Empty, false };
+            yield return new object[] { null, false };
+            yield return new object[] { $@"\\", false };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}{CrossPlatDirSeparatorChar}", false };
+            yield return new object[] { $@"{CrossPlatDirSeparatorChar}{CrossPlatDirSeparatorChar}directory", true };
+        }
+
+        // ReSharper restore HeapView.BoxingAllocation
+        #endregion
+
         #region Method Tests
         [Theory]
         [InlineData('x', true)]
@@ -151,6 +246,61 @@ namespace VelaptorTests
         {
             // Act
             var actual = value.GetLastDirName();
+
+            // Assert
+            Assert.Equal(expected, actual);
+        }
+
+        [Theory]
+        [MemberData(nameof(HasValidFullFilePathSyntaxTestData))]
+        public void HasValidFullFilePathSyntax_WhenInvoked_ReturnsCorrectResult(string path, bool expected)
+        {
+            // Act
+            var actual = path.HasValidFullFilePathSyntax();
+
+            // Assert
+            Assert.Equal(expected, actual);
+        }
+
+        [Theory]
+        [MemberData(nameof(IsInvalidFilePathTestData))]
+        public void HasInvalidFullFilePathSyntax_WhenInvoked_ReturnsCorrectResult(string path, bool expected)
+        {
+            // Act
+            var actual = path.HasInvalidFullFilePathSyntax();
+
+            // Assert
+            Assert.Equal(expected, actual);
+        }
+
+        [Theory]
+        [MemberData(nameof(ContainsValidDriveTestData))]
+        public void HasValidDriveSyntax_WhenInvoked_ReturnsCorrectResult(string dirPath, bool expected)
+        {
+            // Act
+            var actual = dirPath.HasValidDriveSyntax();
+
+            // Assert
+            Assert.Equal(expected, actual);
+        }
+
+        [Theory]
+        [MemberData(nameof(IsFullyQualifiedDirPathTestData))]
+        public void HasValidFullDirPathSyntax_WhenInvoked_ReturnsCorrectResult(string dirPath, bool expected)
+        {
+            // Act
+            var actual = dirPath.HasValidFullDirPathSyntax();
+
+            // Assert
+            Assert.Equal(expected, actual);
+        }
+
+        [Theory]
+        [MemberData(nameof(IsUNCPathTestData))]
+        public void HasValidUNCPathSyntax_WhenInvoked_ReturnsCorrectResult(string path, bool expected)
+        {
+            // Act
+            var actual = path.HasValidUNCPathSyntax();
 
             // Assert
             Assert.Equal(expected, actual);
